fixing test tools check
